--- a/crates/core/src/optimizer/hoist_files.rs
+++ b/crates/core/src/optimizer/hoist_files.rs
@@ -226,7 +226,11 @@ impl<Q: QueryContext> FilenamePatternExtractor<Q> for Predicate<Q> {
                             match &m.pattern {
                                 Some(pattern) => {
                                     // This is the key line of this entire file
-                                    return Ok(Some(pattern.clone()));
+                                    if is_safe_to_hoist(pattern)? {
+                                        return Ok(Some(pattern.clone()));
+                                    } else {
+                                        return Ok(None);
+                                    }
                                 }
                                 None => {}
                             }
@@ -279,5 +283,18 @@ impl<Q: QueryContext> FilenamePatternExtractor<Q> for Predicate<Q> {
             // These are more complicated, implement carefully
             Predicate::Call(_) | Predicate::Not(_) | Predicate::Equal(_) => Ok(None),
         }
+    }
+}
+
+// Check if a filename pattern is safe to hoist.
+// This is not a great implementation, but it's a start.
+// I think a better approach will actually be to introduce a Pattern::FailOpen idea where if any errors are encountered when resolving
+// a pattern, we can just assume it's true. This will allow us to hoist more patterns without worrying about unbound variables.
+fn is_safe_to_hoist<Q: QueryContext>(pattern: &Pattern<Q>) -> Result<bool> {
+    match pattern {
+        Pattern::Includes(inc) => is_safe_to_hoist(&inc.includes),
+        Pattern::StringConstant(_) => Ok(true),
+        // This is conservative, but it's a start
+        _ => Ok(false),
     }
 }