… crates/cli_bin/benchmark/bench.sh
… crates/core/src/optimizer/mod.rs
… crates/core/src/pattern_compiler/auto_wrap.rs
… crates/core/src/test_files.rs
… crates/core/src/test_utils.rs
… crates/grit-pattern-matcher/src/pattern/any.rs
… crates/grit-pattern-matcher/src/pattern/includes.rs
… crates/grit-pattern-matcher/src/pattern/variable.rs
… crates/grit-pattern-matcher/src/pattern/within.rs