PoC for hoisting filename matches so we avoid parsing unnecessary files for patterns like this:

```
`foo` where {
  $filename <: includes "nice.js"
}
```

## Impact

Even on small repos it makes a large difference.

before:
```
Processed 666 files and found 1 matches
  Time (mean ± σ):     835.1 ms ± 824.5 ms    [User: 4106.8 ms, System: 251.8 ms]
  Range (min … max):   548.3 ms … 3181.0 ms    10 runs
```

after:
```
Processed 666 files and found 1 matches
  Time (mean ± σ):     157.0 ms ±  14.1 ms    [User: 275.6 ms, System: 119.8 ms]
  Range (min … max):   147.7 ms … 204.1 ms    14 runs
```

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit


- **New Features**
  - Introduced filename pattern extraction functionality to optimize file matching in queries.
  - Added new test scenarios for filename query optimization.

- **Bug Fixes**
  - Modified behavior in test utilities to ensure proper handling of unexpected file reads.

- **Improvements**
  - Enhanced pattern matching by optimizing filename patterns for better performance.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->